requirements.txt
torch>=2.0.0
numpy
opencv-python
pillow
tqdm
albumentations
segmentation-models-pytorch
scikit-learn
matplotlib